Renamed frinal score calculator
