+++ src/main/frontend/front_end_api.py
@@ -9,7 +9,7 @@ from main.data_types.sbom_types.sbom import Sbom
 from main.data_types.user_requirements import UserRequirements
 from main.data_types.event import Event
 from main.sbom_processor import SbomProcessor, SbomProcessorStatus
-from main.frontend.final_score_calculator import grade_dependencies
+from main.frontend.dependency_grader import grade_dependencies
 
 
 class FrontEndAPI:
